feat: baremetal pkg manager
