+++ nvim/.config/nvim/init.lua
@@ -4,57 +4,19 @@
 vim.g.mapleader = ' '
 vim.g.maplocalleader = ' '
 
--- trae stuff
+-- trae
 vim.g.trae_disable_autocompletion = true
 vim.g.trae_no_map_tab = true
 vim.g.trae_disable_bindings = true
 
 -- set transparent
-vim.api.nvim_set_hl(0, "Normal", { bg = "none" })
-vim.api.nvim_set_hl(0, "NonText", { bg = "none" })
+vim.api.nvim_set_hl(0, 'Normal', { bg = 'none' })
+vim.api.nvim_set_hl(0, 'NonText', { bg = 'none' })
 
 -- NOTES: experimental
-require('vim._core.ui2').enable({})
+require('vim._core.ui2').enable {}
 
--- [[ Install `lazy.nvim` plugin manager ]]
---    https://github.com/folke/lazy.nvim
---    `:help lazy.nvim.txt` for more info
-local lazypath = vim.fn.stdpath 'data' .. '/lazy/lazy.nvim'
-if not vim.loop.fs_stat(lazypath) then
-  vim.fn.system {
-    'git',
-    'clone',
-    '--filter=blob:none',
-    'https://github.com/folke/lazy.nvim.git',
-    '--branch=stable', -- latest stable release
-    lazypath,
-  }
-end
-vim.opt.rtp:prepend(lazypath)
-
-
--- [[ Configure plugins ]]
-require('lazy').setup({
-  import = 'custom/plugins',
-}, {
-  ui = { border = 'rounded' },
-  change_detection = {
-    notify = false,
-  },
-  performance = {
-    rtp = {
-      disabled_plugins = {
-        'gzip',
-        'netrwPlugin',
-        'rplugin',
-        'tarPlugin',
-        'tohtml',
-        'tutor',
-        'zipPlugin',
-      },
-    },
-  },
-})
+require 'custom_plugins'
 
 -- The line beneath this is called `modeline`. See `:help modeline`
 -- vim: ts=2 sts=2 sw=2 et
