--- nvim/.config/nvim/lua/custom/plugins/arrow.lua
@@ -1,9 +1,0 @@
-return {
-  'otavioschwanck/arrow.nvim',
-  event = 'VeryLazy',
-  opts = {
-    show_icons = false,
-    leader_key = '<space>;',
-    buffer_leader_key = '<space>m',
-  },
-}
