--- nvim/.config/nvim/lua/custom/plugins/grug.lua
@@ -1,5 +1,0 @@
-return {
-  'MagicDuck/grug-far.nvim',
-  cmd = 'GrugFar',
-  opts = {},
-}
