--- nvim/.config/nvim/lua/custom/plugins/init.lua
@@ -1,8 +1,0 @@
-return {
-  { 'tpope/vim-sleuth', event = 'VeryLazy' },
-  {
-    'saghen/blink.indent',
-    event = 'VeryLazy',
-    opts = {},
-  }
-}
