--- nvim/.config/nvim/lua/custom/plugins/outline.lua
@@ -1,13 +1,0 @@
-return {
-  'hedyhli/outline.nvim',
-  lazy = true,
-  cmd = { 'Outline', 'OutlineOpen' },
-  keys = {
-    { '<leader>o', '<cmd>Outline<CR>', desc = 'Toggle outline' },
-  },
-  opts = {
-    outline_window = {
-      position = 'left',
-    },
-  },
-}
