--- nvim/.config/nvim/lua/custom/plugins/trouble.lua
@@ -1,12 +1,0 @@
-return {
-  'folke/trouble.nvim',
-  opts = {},
-  cmd = 'Trouble',
-  keys = {
-    {
-      '<leader>tt',
-      '<cmd>Trouble diagnostics toggle<cr>',
-      desc = 'Diagnostics (Trouble)',
-    },
-  },
-}
