--- nvim/.config/nvim/lua/custom/plugins/undotree.lua
@@ -1,6 +1,0 @@
-return {
-  'mbbill/undotree',
-  keys = {
-    { '<leader>u', '<cmd>UndotreeToggle<cr>', desc = 'Undo tree toggle' },
-  },
-}
