--- nvim/.config/nvim/lua/custom/plugins/whichkey.lua
@@ -1,16 +1,0 @@
-return {
-    'folke/which-key.nvim',
-    event = 'VeryLazy',
-    opts = {
-        preset = 'helix',
-    },
-    keys = {
-        {
-            '<leader>?',
-            function()
-                require('which-key').show { global = false }
-            end,
-            desc = 'Buffer Local Keymaps (which-key)',
-        },
-    },
-}
