+++ nvim/.config/nvim/lua/custom_plugins/lsp.lua
@@ -1,0 +1,56 @@
+vim.pack.add({
+  gh 'neovim/nvim-lspconfig',
+  gh 'williamboman/mason.nvim',
+  gh 'mason-org/mason-lspconfig.nvim',
+}, { confirm = false })
+
+local mason = require 'mason'
+mason.setup {
+  ensure_installed = {
+    'cssls',
+    'eslint',
+    'gopls',
+    'lua_ls',
+    'oxlint',
+    'rust_analyzer',
+    'stylua',
+    'tailwindcss',
+    'vtsls',
+  },
+}
+
+local capabilities = require('blink.cmp').get_lsp_capabilities()
+
+vim.lsp.config('vtsls', {
+  settings = require('lsp_config.vtsls').settings,
+})
+
+vim.lsp.config('lua_ls', {
+  settings = require('lsp_config.lua_ls').settings,
+})
+
+local eslint_lsp = require 'lsp_config.eslint'
+vim.lsp.config('eslint', {
+  cmd = eslint_lsp.cmd,
+  settings = eslint_lsp.settings,
+})
+
+vim.lsp.config('gopls', {
+  settings = require('lsp_config.gopls').settings,
+})
+
+vim.lsp.config('*', {
+  capabilities = capabilities,
+})
+
+vim.lsp.enable {
+  'copilot',
+  'cssls',
+  'eslint',
+  'gopls',
+  'lua_ls',
+  'oxlint',
+  'rust_analyzer',
+  'tailwindcss',
+  'vtsls',
+}
